+++ cider-repl.el
@@ -363,6 +363,16 @@ This will not work on non-current prompts."
 
 (put 'cider-save-marker 'lisp-indent-function 1)
 
+(defun cider-repl-default-prompt (namespace)
+  "Return a prompt string that mentions NAMESPACE."
+  (format "%s> " namespace))
+
+(defcustom cider-repl-prompt-function #'cider-repl-default-prompt
+  "A function that returns a prompt string.
+Takes one argument, a namespace name."
+  :type 'function
+  :group 'cider-repl)
+
 (defun cider-repl--insert-prompt (namespace)
   "Insert the prompt (before markers!), taking into account NAMESPACE.
 Set point after the prompt.
@@ -372,7 +382,7 @@ Return the position of the prompt beginning."
     (cider-save-marker cider-repl-output-end
       (unless (bolp) (insert-before-markers "\n"))
       (let ((prompt-start (point))
-            (prompt (format "%s> " namespace)))
+            (prompt (funcall cider-repl-prompt-function namespace)))
         (cider-propertize-region
             '(font-lock-face cider-repl-prompt-face read-only t intangible t
                              field cider-repl-prompt